+++ changelog.md
@@ -3,3 +3,6 @@
 
 #### 24/05/2025 19:27
 - Fix condition
+
+#### 24/05/2025 20:08
+- Configuration based time window